Fix #1547
